backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asyncpg==0.31.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
curl_cffi==0.13.0
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.135.1
frozendict==2.4.7
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.4.3
openai==2.28.0
packaging==26.0
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.4
protobuf==7.34.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.2
yfinance==1.2.0
